backend/requirements/base.txt
Django==5.0.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
django-cors-headers==4.3.1
python-decouple==3.8
psycopg2-binary==2.9.9
celery==5.3.4
redis==5.0.1
boto3==1.34.14
django-storages==1.14.2
Pillow==10.1.0
langchain==0.1.0
openai==1.7.1
reportlab==4.0.8
markdown==3.5.1
gunicorn==21.2.0
python-jose==3.3.0
drf-spectacular==0.27.0